+++ test/nextgen/reference/compare.cpp
@@ -1,5 +1,6 @@
 //
 // SPDX-FileCopyrightText: Copyright 2025-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2026 Meta Platforms, Inc. and affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -14,6 +15,7 @@
 #include <ostream>
 
 #include "test/common/assert.hpp"
+#include "test/common/bfloat16.hpp"
 #include "test/common/data_type.hpp"
 #include "test/common/float16.hpp"
 #include "test/common/int2.hpp"
@@ -103,6 +105,9 @@ CompareFn make_compare_plain_2d(DataType dtype) {
 
         case DataType::FP16:
             return compare_plain_2d<Float16>;
+
+        case DataType::BF16:
+            return compare_plain_2d<BFloat16<>>;
 
         case DataType::I32:
             return compare_plain_2d<int32_t>;