+++ test/nextgen/reference/print.cpp
@@ -1,5 +1,6 @@
 //
 // SPDX-FileCopyrightText: Copyright 2025-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2026 Meta Platforms, Inc. and affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -14,6 +15,7 @@
 #include <string>
 
 #include "test/common/assert.hpp"
+#include "test/common/bfloat16.hpp"
 #include "test/common/data_type.hpp"
 #include "test/common/float16.hpp"
 #include "test/common/int2.hpp"
@@ -70,6 +72,9 @@ PrintFn make_print_array(DataType dtype) {
 
         case DataType::FP16:
             return print_array<Float16>;
+
+        case DataType::BF16:
+            return print_array<BFloat16<>>;
 
         case DataType::I32:
             return print_array<int32_t>;